Move `ColorConfig` into `color.rs`
